indexing integer debug + removed .o .so .c

+++ orb/utils/image.py
@@ -1025,8 +1025,8 @@ def fit_calibration_laser_map(calib_laser_map, calib_laser_nm, pixel_size=15.,
     cx = calib_laser_map.shape[0]/2.
     cy = calib_laser_map.shape[1]/2.
     center_calib_nm = np.nanmean(
-        calib_laser_map[int(cx-0.5):math.ceil(cx-0.5+1),
-                        int(cy-0.5):math.ceil(cy-0.5+1)])
+        calib_laser_map[int(cx-0.5):int(math.ceil(cx-0.5+1)),
+                        int(cy-0.5):int(math.ceil(cy-0.5+1))])
     
     theta_c = math.acos(calib_laser_nm/center_calib_nm) / math.pi * 180.
     print 'Angle at the center of the frame: {}'.format(theta_c)
